+++ CHANGELOG.md
@@ -5,7 +5,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Ideas/Upcoming]
 Pending changes
-NavigationPresenter UI element caching
 Switch Animations
 
 ## [Unreleased]
@@ -13,6 +12,7 @@ Switch Animations
 * UIDispatcher in Utils to have an easy access to the UIDispatcher in ViewModels
 * ColorPicker
 * FontPicker
+* Support control caching in the NavigationPresenter
 
 ## [1.0.0] - 2021-06-13
 ### Added